Refactor Mopidy classes into their own source files.
