+++ tests/conftest.py
@@ -10,7 +10,7 @@ import pytest
 
 import requests
 
-from mopidy_pandora import actor
+from mopidy_pandora import backend
 
 MOCK_STATION_SCHEME = "station"
 MOCK_STATION_NAME = "Mock Station"
@@ -52,7 +52,7 @@ def config():
 
 
 def get_backend(config, simulate_request_exceptions=False):
-    obj = actor.PandoraBackend(config=config, audio=None)
+    obj = backend.PandoraBackend(config=config, audio=None)
 
     if simulate_request_exceptions:
         type(obj.api.transport).__call__ = transport_call_request_exception_mock
